+++ CHANGELOG.md
@@ -10,8 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `PFAnimation` class that does quintic bezier interpolation between keyframes.
 - `PFEditor` class that allows editing keyframes.
-- Custom handlers for editing parameters.
-- Export and import.
+    - Custom handlers for editing parameters.
+    - 1-9 selects the current handler.
+    - Return enters pointer lock mode to edit parameter values.
+    - Backspace deletes the current keyframe.
+- Export and import by pressing `E` and `I` respectively.
 
 ### Changed
 - Renamed `examples/index.html` to `examples/playback.html`.